TASK:1416 Package name contains upper case characters
