+++ jdi-light-bdd-tests/src/test/java/cucmbertests/stepdefs/UserStepdefs.java
@@ -1,4 +1,4 @@
-package cucmberTests.stepdefs;
+package cucmbertests.stepdefs;
 
 import com.epam.jdi.light.elements.common.UIElement;
 import com.epam.jdi.light.elements.composite.WebPage;
